Remove STM32 MCU-family ifdefs from src/main
Replace MCU-family guards with capability flags defined in
target/platform headers:

- ENABLE_SDIO_PIN_CONFIG gates the sdioPinConfig_t PG (pg/sdio.c),
  the CLI resource sdio_* commands (cli/cli.c), and the sdio_device
  setting (cli/settings.c). Set for H5/H7/N6 targets. Previously
  sdioPinConfig_t was registered on H5/H7/N6 but the CLI resources
  and sdio_device setting were H7-only.

- USE_TIMER_UP_CONFIG (existing flag) gates the TIMUP DMA CLI
  resource (cli/cli.c) instead of STM32H7/G4, extending it to N6
  which already defines USE_TIMER_UP_CONFIG.

- ENABLE_SDIO_EXTERNAL_DMA gates external DMA channel allocation
  in sdcard_sdio_baremetal.c. Set for STM32F4/F7 (STM32 platform.h)
  and APM32 (APM32 platform.h). H5/H7/N6 SDMMC peripherals use
  internal DMA and do not need allocation.

- fc/hardfaults.c split. fc/faults.c holds the generic
  systemFaultAction() (stop motors, LEDs, transponder).
  src/platform/common/stm32/fault_handlers.c holds the ARM
  Cortex-M HardFault_Handler and DEBUG_HARDFAULTS trampoline
  target, included by STM32/AT32/APM32 platforms. MemManage_Handler
  removed: no code in the tree enables MemManage fault (no
  SHCSR.MEMFAULTENA is ever set), so the handler was dead on every
  target.
