+++ src/platform/APM32/include/platform/platform.h
@@ -127,6 +127,10 @@
 #define USE_PERSISTENT_OBJECTS
 #define USE_LATE_TASK_STATISTICS
 
+#if !defined(ENABLE_SDIO_EXTERNAL_DMA)
+#define ENABLE_SDIO_EXTERNAL_DMA 1
+#endif
+
 #define USE_OVERCLOCK
 #define ENABLE_OVERCLOCK_192_MHZ 1
 #define ENABLE_OVERCLOCK_216_MHZ 1
