+++ src/platform/AT32/mk/AT32F4.mk
@@ -142,7 +142,8 @@ MCU_COMMON_SRC = \
             common/stm32/ledstrip_ws2811_stm32.c \
             common/stm32/debug_pin.c \
             common/stm32/adc_impl.c \
-            common/stm32/expresslrs_driver_hw.c
+            common/stm32/expresslrs_driver_hw.c \
+            common/stm32/fault_handlers.c
 
 SPEED_OPTIMISED_SRC += \
             common/stm32/dshot_bitbang_shared.c \
