+++ src/platform/STM32/mk/STM32_COMMON.mk
@@ -23,7 +23,8 @@ MCU_COMMON_SRC += \
             common/stm32/ledstrip_ws2811_stm32.c \
             common/stm32/debug_pin.c \
             common/stm32/adc_impl.c \
-            common/stm32/expresslrs_driver_hw.c
+            common/stm32/expresslrs_driver_hw.c \
+            common/stm32/fault_handlers.c
 
 SIZE_OPTIMISED_SRC += \
             drivers/bus_spi_config.c \
